+++ migrations/mysql/20260414000000_torrust_user_token_generation.sql
@@ -1,0 +1,1 @@
+ALTER TABLE torrust_users ADD COLUMN token_generation BIGINT NOT NULL DEFAULT 0