+++ src/web/api/server/v1/extractors/user_id.rs
@@ -4,10 +4,9 @@ use axum::extract::{FromRef, FromRequestParts};
 use axum::http::request::Parts;
 use axum::response::{IntoResponse, Response};
 
-use super::bearer_token;
 use crate::common::AppData;
-use crate::errors::AuthError;
 use crate::models::user::UserId;
+use crate::web::api::server::v1::extractors::bearer_token::BearerToken;
 
 pub struct ExtractLoggedInUser(pub UserId);
 
@@ -19,20 +18,15 @@ where
     type Rejection = Response;
 
     async fn from_request_parts(parts: &mut Parts, state: &S) -> Result<Self, Self::Rejection> {
-        let maybe_bearer_token = match bearer_token::Extract::from_request_parts(parts, state).await {
-            Ok(maybe_bearer_token) => maybe_bearer_token.0,
-            Err(_) => return Err(AuthError::TokenNotFound.into_response()),
-        };
+        let token = BearerToken::from_request_parts(parts, state).await?;
 
-        //Extracts the app state
         let app_data = Arc::from_ref(state);
 
-        #[allow(clippy::option_if_let_else)]
-        let result = match app_data.auth.get_user_id_from_bearer_token(maybe_bearer_token) {
-            Ok(user_id) => Ok(Self(user_id)),
-            Err(_) => Err(AuthError::LoggedInUserNotFound.into_response()),
-        };
-
-        result
+        app_data
+            .auth
+            .get_user_id_from_bearer_token(token)
+            .await
+            .map(Self)
+            .map_err(IntoResponse::into_response)
     }
 }